Fix flaky test
The test "POSTs the visit to the backend" was dependent on the browser running
`fauxJax.once...` before the call to `ahoy.start()` from the previous test
made Ahoy POST the visit to the browser.

This approach worked on some browsers, but failed consistently on others (IE11
for instance on SauceLabs).

Since we can only start Ahoy once, I moved the expectations regarding the visit
creation to the same test that initializes Ahoy, so we won't depend on async
behavior between tests.